Update LCA input to use key/value
This is an attempt at consolidating the user input for new and existing
LCA types, opting for a single, key/value input space rather than
separate inputs `LCA_Type` and `LCA_Data`. This reduces the amount of
checks. It also makes the input require key/value for metrics, to
prevent possible data entry mistakes due to ordering confusion. We'll
need to make sure that the documentation is sufficient for new users of
this feature.